修改float为cgfloat
修改float为cgfloat

+++ UI/src/scrollableListController/model/MSSectionViewModel.h
@@ -11,7 +11,7 @@
 @protocol MSSectionViewModel <NSObject>
 
 
-@property (nonatomic, assign) float height;
+@property (nonatomic, assign) CGFloat height;
 
 
 @property (nonatomic, strong) Class Class;
